ml_service/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
pydantic
requests
python-dotenv